refactor(editor): extract extensions array into build-extensions.ts (issue #1921)
Extract the 196-line extensions array (Editor.svelte:441-637) into a new
pure function buildExtensions() in src/renderer/lib/editor/build-extensions.ts.
This seals a major architectural seam and makes the extension set unit-testable
without mounting the component (issue requirement).

The extracted function captures all closure-dependent state via a
BuildExtensionsOptions interface that includes:
- Compartments for live reconfiguration (theme, font size, settings)
- Props callbacks (onNavigate, onOpenSource, onRunCell, etc.)
- Accessors for Editor-owned state (getSavedSelection, setSavedSelection)
- Handler callbacks (showContextMenu, onImageDrop)

Editor.svelte now composes the array via a single function call, dropping
196 lines of complex extension wiring. The behavior is identical — the change
is purely architectural.

Editor.svelte shrinks from 1208 → 1024 lines (184-line reduction).
build-extensions.ts: 256 lines (under 600-line threshold, no budget entry).

- pnpm lint: passes (tsc, svelte-check, eslint)
- pnpm test: all 625 test files pass
- File budgets updated in same diff

Related: #1903 (broader CodeMirror extraction), #1854 (file-size budgets)
